move tests for ValidationScheduler to a test file for that class
